[Refactor] Modify addListener method to contructor with rename as initialize
